refactor: Update AccessLogPage and AccessRequestPage to use "overflow-scroll" class
The AccessLogPage and AccessRequestPage components were modified to use the "overflow-scroll" class instead of "overflow-y-scroll" in the root div. This change ensures that the page content is scrollable in both the vertical and horizontal directions when there is overflow. The "overflow-scroll" class provides a more flexible scrolling behavior compared to the previous "overflow-y-scroll" class.